backend/requirements.txt
streamlit
flask
pymupdf
python-docx
sentence-transformers
faiss-cpu
fuzzywuzzy[speedup]
spacy
sqlalchemy
torch 